fix cpu and gpu execution mismatch.

+++ src/targets/gpu/device/convert.cpp
@@ -23,7 +23,7 @@ void convert(hipStream_t stream,
                     float res  = input_ptr[i] * scale + shift;
                     int factor = (res > 0) ? 1 : -1;
                     output_ptr[i] =
-                        std::min<int8_t>(std::max<float>(-128, res + factor * 0.5), 127);
+                        static_cast<int8_t>(std::min<float>(std::max<float>(-128.0f, res + factor * 0.5), 127.0f));
                 });
             }
             else
